+++ packages/library/src/index.ts
@@ -6,7 +6,7 @@ export { Components, JSX } from './components'
 /**
  * Types
  */
-export { BalOptionValue } from './components/bal-select-option/bal-select-option.type'
+export { BalOptionValue, NewBalOptionValue } from './components/bal-select-option/bal-select-option.type'
 export { BalCardStepOption } from './components/bal-card-step/bal-card-step.type'
 export { BalTabOption } from './components/bal-tabs/bal-tab.type'
 export { FileUploadRejectionReason, FileUploadRejectedFile } from './components/bal-file-upload/bal-file-upload.type'