+++ packages/testing/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.14.2](https://github.com/baloise/ui-library-next/compare/v1.14.1...v1.14.2) (2020-11-10)
+
+**Note:** Version bump only for package @baloise/ui-library-next-testing
+
+
+
+
+
+## [1.14.1](https://github.com/baloise/ui-library-next/compare/v1.14.0...v1.14.1) (2020-11-10)
+
+**Note:** Version bump only for package @baloise/ui-library-next-testing
+
+
+
+
+
+# [1.14.0](https://github.com/baloise/ui-library-next/compare/v1.13.2...v1.14.0) (2020-11-09)
+
+**Note:** Version bump only for package @baloise/ui-library-next-testing
+
+
+
+
+
+## [1.13.2](https://github.com/baloise/ui-library-next/compare/v1.13.1...v1.13.2) (2020-11-09)
+
+**Note:** Version bump only for package @baloise/ui-library-next-testing
+
+
+
+
+
 ## [1.13.1](https://github.com/baloise/ui-library-next/compare/v1.13.0...v1.13.1) (2020-11-09)
 
 **Note:** Version bump only for package @baloise/ui-library-next-testing